Modify FFSfile internals to support larger-than-32-bit record lengths.  For the actual data files, the limit is now 48 bits max record length.
git-svn-id: http://svn.research.cc.gatech.edu/kaos/ffs/trunk@8273 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ ffs/ffs_internal.h
@@ -192,7 +192,7 @@ make_tmp_buffer(FFSBuffer buf, int size);
 
 extern int
 FFS_decode_length_format(FFSContext context, FFSTypeHandle ioformat, 
-			 int record_length);
+			 long record_length);
 
 #define Max(i,j) ((i<j) ? j : i)
